fix(account): hash comparison triggered on missing session
This shouldn't have been happened. When Valkey doesn't find the session
that was requested via `hgetall`, it returns an empty object. Previously,
there was not any guard to handle this and the `Session` domain object
creation with this data somehow succeeded. This had caused an error being
thrown during hash comparison because the digest value was `undefined`