feat(game): embed game stats to player connection
Game stats could be used for matchmaking purpose or any other possible
usecases in the future. This is possible thanks to the extension of the
`account` upstream API.

This commit also refactors a portion of the `wsApp`. The main `game`
module now responsible on determining the implementation of commands
that should to be done when a player had disconnected, such as clean-up
operations.