ref: generalize errors
Previously error definitions are highly bounded to HTTP errors. Now that
the repo contains services that uses Websocket, using the same approach
would cause duplications.

So, instead of handling the transformation of errors to something that
could be sent utilizing the protocol, the transformation is handled by
adapters. This way, modules that needs to the errors in their own unique
format could simply implement their own adapter.