feat(sys/lib): provide connection look up using id
Previously, `WsConnectionManager` only provides functionality to send
something by their connection name. This might be problematic if a part
of the program doesn't have the connection name. Fortunately, most of
them knows a way of identifying who's who by using some kind of
identifier, usually player's id. A method to do a look up for connection
a name using it may come in handy in the future