feat(game): limit board state visibility for players
While the client (front-end) could decide what and what not to show for
the player, it is still possible for a player to "intercept" the response
received to see the actual payload before it get filtered by the client.
If they could see what they shouldn't see (for example, their opponent's
face-down card), the fairness of the game would be compromised.