fix(game): unhandled stale data
When the server crashes/restarted, all of the match data would be gone.
However, the records of which players joins which match still persist
in the cache, albeit being stale. Without some "housekeeping" of these
stale data, they would always be there and prevents player to join the
game queue.