fix(game): wrong game flow when ending a round
The basic rule is that a round would be concluded after both players
consecutively 'passed' (or "stay" in the original game). However, if
one of the players use any trump cards before the pass, the opponent
should be given a chance to retaliate (either by doing another hit
or use trump cards). This applies vice versa, until the last player
decided to pass (chose not to retaliate).

The previous implementation didn't consider the trump card being used
before a player passed. It means that the opponent of the last player
that had both passed and used trump cards wasn't given any chance to
retaliate.